apply requested changes
