+++ node/middlewares/internals/brand.ts
@@ -1,66 +1,50 @@
 import { Brand } from 'vtex.catalog-graphql'
-import { InternalInput } from 'vtex.rewriter'
 
 import { Context } from '../../typings/global'
+import { filterStoreBindings } from '../../utils/bindings'
 import {
-  filterBindings,
-  getPath,
   INDEXED_ORIGIN,
   PAGE_TYPES,
-  slugify,
+  routeFormatter,
   STORE_LOCATOR,
-} from './utils'
-
-const getBrandInternals = (
-  path: string,
-  id: string,
-  bindings: string[]
-): InternalInput[] =>
-  bindings.map(binding => ({
-    binding,
-    declarer: STORE_LOCATOR,
-    from: path,
-    id,
-    origin: INDEXED_ORIGIN,
-    query: {
-      map: 'b',
-    },
-    type: PAGE_TYPES.BRAND,
-  }))
-
-const getBrandNotFoundInternals = (
-  path: string,
-  bindings: string[]
-): InternalInput[] =>
-  bindings.map(binding => ({
-    binding,
-    declarer: STORE_LOCATOR,
-    from: path,
-    id: 'brand',
-    origin: INDEXED_ORIGIN,
-    type: PAGE_TYPES.SEARCH_NOT_FOUND,
-  }))
+} from '../../utils/internals'
+import { slugify } from '../../utils/slugify'
 
 export async function brandInternals(ctx: Context, next: () => Promise<void>) {
   const {
     clients: { apps },
-    vtex: { logger },
     state,
   } = ctx
-  try {
-    const bindings = filterBindings(ctx.state.tenantInfo, null)
-    const brand: Brand = ctx.body
-    const brandName = slugify(brand.name)
-    const path = await getPath(PAGE_TYPES.BRAND, { brand: brandName }, apps)
+  const brand: Brand = ctx.body
+  const { name, active, id } = brand
+  const brandName = slugify(name)
+  const bindings = filterStoreBindings(ctx.state.tenantInfo)
 
-    const internals = brand.active
-      ? getBrandInternals(path, brand.id, bindings)
-      : getBrandNotFoundInternals(path, bindings)
-
-    state.internals = internals
-  } catch (error) {
-    logger.error(error)
+  if (bindings.length === 0) {
+    return
   }
+
+  const formatRoute = await routeFormatter(apps, PAGE_TYPES.BRAND)
+
+  const internals = await Promise.all(
+    bindings.map(async binding => {
+      const { defaultLocale, id: bindingId } = binding
+      const translated = brandName
+      const path = formatRoute({ brand: translated })
+
+      return {
+        binding: bindingId,
+        declarer: STORE_LOCATOR,
+        from: path,
+        id,
+        origin: INDEXED_ORIGIN,
+        query: active ? { map: 'b' } : null,
+        type: active ? PAGE_TYPES.BRAND : PAGE_TYPES.SEARCH_NOT_FOUND,
+      }
+    })
+  )
+
+  state.internals = internals
 
   await next()
 }
