+++ node/middlewares/internals/product.ts
@@ -1,38 +1,13 @@
 import { Product } from 'vtex.catalog-graphql'
-import { InternalInput } from 'vtex.rewriter'
 
 import { Context } from '../../typings/global'
+import { filterBindingsBySalesChannel } from '../../utils/bindings'
 import {
-  filterBindings,
-  getPath,
   INDEXED_ORIGIN,
   PAGE_TYPES,
+  routeFormatter,
   STORE_LOCATOR,
-} from './utils'
-
-const getProductInternal = (
-  path: string,
-  id: string,
-  bindings: string[]
-): InternalInput[] =>
-  bindings.map(binding => ({
-    binding,
-    declarer: STORE_LOCATOR,
-    from: path,
-    id,
-    origin: INDEXED_ORIGIN,
-    type: PAGE_TYPES.PRODUCT,
-  }))
-
-const getProductNotFoundInternal = (path: string): InternalInput[] => [
-  {
-    declarer: STORE_LOCATOR,
-    from: path,
-    id: 'product',
-    origin: INDEXED_ORIGIN,
-    type: PAGE_TYPES.PRODUCT_NOT_FOUND,
-  },
-]
+} from '../../utils/internals'
 
 export async function productInternals(
   ctx: Context,
@@ -40,24 +15,40 @@ export async function productInternals(
 ) {
   const {
     clients: { apps },
-    vtex: { logger },
     state,
   } = ctx
-  try {
-    const product: Product = ctx.body
-    const slug = product.linkId?.toLowerCase()
-    const path = await getPath(PAGE_TYPES.PRODUCT, { slug }, apps)
-    const bindings = filterBindings(ctx.state.tenantInfo, product.salesChannel)
+  const product: Product = ctx.body
+  const { linkId, isActive, id } = product
+  const slug = linkId?.toLowerCase()
+  const bindings = filterBindingsBySalesChannel(
+    ctx.state.tenantInfo,
+    product.salesChannel
+  )
 
-    const internals =
-      product.isActive && bindings.length > 0
-        ? getProductInternal(path, product.id, bindings)
-        : getProductNotFoundInternal(path)
-
-    state.internals = internals
-  } catch (error) {
-    logger.error(error)
+  if (!slug || bindings.length === 0) {
+    return
   }
+
+  const formatRoute = await routeFormatter(apps, PAGE_TYPES.PRODUCT)
+
+  const internals = await Promise.all(
+    bindings.map(async binding => {
+      const { defaultLocale, id: bindingId } = binding
+      const translated = slug
+      const path = formatRoute({ slug: translated })
+
+      return {
+        binding: bindingId,
+        declarer: STORE_LOCATOR,
+        from: path,
+        id,
+        origin: INDEXED_ORIGIN,
+        type: isActive ? PAGE_TYPES.PRODUCT : PAGE_TYPES.PRODUCT_NOT_FOUND,
+      }
+    })
+  )
+
+  state.internals = internals
 
   await next()
 }
