+++ node/middlewares/internals/save.ts
@@ -4,20 +4,15 @@ import { Context } from '../../typings/global'
 
 export async function saveInternals(ctx: Context, next: () => Promise<void>) {
   const {
-    vtex: { logger },
     clients: { rewriter },
     state: { internals },
   } = ctx
-  try {
-    if (internals) {
-      if (LINKED) {
-        // eslint-disable-next-line no-console
-        console.log('Saving', internals)
-      }
-      await rewriter.saveManyInternals(internals)
+  if (internals) {
+    if (LINKED) {
+      // eslint-disable-next-line no-console
+      console.log('Saving', internals)
     }
-  } catch (error) {
-    logger.error(error)
+    await rewriter.saveManyInternals(internals)
   }
 
   await next()
