+++ node/utils/slugify.ts
@@ -1,0 +1,8 @@
+/* eslint-disable no-useless-escape */
+
+export const slugify = (str: string) =>
+  str
+    .toLowerCase()
+    .normalize('NFD')
+    .replace(/[\u0300-\u036f]/g, '')
+    .replace(/[*+~.()'"!:@&\[\]`,/ %$#?{}|><=_^]/g, '-')
